# apps/api/src/modules/results/results.controller.ts
import {
  Body,
  Controller,
  Get,
  Param,
  Post,
  Query,
  Res,
} from '@nestjs/common';
import { ApiTags } from '@nestjs/swagger';
import { Response } from 'express';
import { Role } from '@prisma/client';
import { ResultsService } from './results.service';
import { ComparisonService } from './comparison.service';
import {
  AggregateResultDto,
  CompareResultsQuery,
  DistributionQuery,
  ExportResultQuery,
  ListResultsQuery,
  ResultDetailQuery,
  SummaryTableQuery,
} from './dto/result.dto';
import {
  CompareResultDto,
  DistributionResultDto,
  SummaryRowDto,
} from './dto/result-response.dto';
import {
  ApiOkEnvelope,
  ApiOkEnvelopeArray,
} from '../../common/swagger/api-envelope.decorator';
import { Roles } from '../../common/decorators/roles';
import { CurrentUser, AuthUser } from '../../common/decorators/current-user';

const XLSX_MIME =
  'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet';

@ApiTags('results')
@Controller('results')
export class ResultsController {
  constructor(
    private readonly resultsService: ResultsService,
    private readonly comparisonService: ComparisonService,
  ) {}

  @Get()
  list(@CurrentUser() user: AuthUser, @Query() query: ListResultsQuery) {
    return this.resultsService.list(user, query);
  }

  // ── YoY: 연도 누적 비교 (정적 경로 — :userId 보다 먼저 선언) ──
  @Get('compare')
  @ApiOkEnvelope(CompareResultDto)
  compare(@CurrentUser() user: AuthUser, @Query() query: CompareResultsQuery) {
    return this.comparisonService.compare(user, query);
  }

  @Get('distribution')
  @ApiOkEnvelope(DistributionResultDto)
  distribution(@CurrentUser() user: AuthUser, @Query() query: DistributionQuery) {
    return this.comparisonService.distribution(user, query);
  }

  // 평가자정리 표 — 다단계 평가 요약(정적 경로, :userId 보다 먼저).
  @Get('summary')
  @ApiOkEnvelopeArray(SummaryRowDto)
  summary(@CurrentUser() user: AuthUser, @Query() query: SummaryTableQuery) {
    return this.resultsService.summaryTable(user, query);
  }

  @Post('aggregate')
  @Roles(Role.hr_admin)
  aggregate(@CurrentUser() user: AuthUser, @Body() dto: AggregateResultDto) {
    return this.resultsService.aggregate(user, dto);
  }

  // M3 Item 9: 개인 평가 결과 내보내기 (excel | pdf=인쇄용 HTML).
  @Get(':userId/export')
  async export(
    @CurrentUser() user: AuthUser,
    @Param('userId') userId: string,
    @Query() query: ExportResultQuery,
    @Res() res: Response,
  ) {
    const out = await this.resultsService.export(user, userId, query);
    if (out.kind === 'excel') {
      res.setHeader('Content-Type', XLSX_MIME);
      res.setHeader(
        'Content-Disposition',
        `attachment; filename="result-${userId}.xlsx"`,
      );
      res.send(out.buffer);
      return;
    }
    res.setHeader('Content-Type', 'text/html; charset=utf-8');
    res.send(out.html);
  }

  @Get(':userId')
  getDetail(
    @CurrentUser() user: AuthUser,
    @Param('userId') userId: string,
    @Query() query: ResultDetailQuery,
  ) {
    return this.resultsService.getDetail(user, userId, query);
  }
}