# packages/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/results": {
      "get": {
        "operationId": "ResultsController_list",
        "parameters": [
          {
            "name": "cycleId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "results"
        ]
      }
    },
    "/api/v1/results/compare": {
      "get": {
        "operationId": "ResultsController_compare",
        "parameters": [
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cycleIds",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/CompareResultDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "results"
        ]
      }
    },
    "/api/v1/results/distribution": {
      "get": {
        "operationId": "ResultsController_distribution",
        "parameters": [
          {
            "name": "scope",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "group",
                "division",
                "team"
              ],
              "type": "string"
            }
          },
          {
            "name": "deptId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cycleIds",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "legalEntity",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "energyx",
                "mirae_plan"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/DistributionResultDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "results"
        ]
      }
    },
    "/api/v1/results/summary": {
      "get": {
        "operationId": "ResultsController_summary",
        "parameters": [
          {
            "name": "cycleId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/SummaryRowDto"
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "number"
                        },
                        "pageSize": {
                          "type": "number"
                        },
                        "total": {
                          "type": "number"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "results"
        ]
      }
    },
    "/api/v1/results/aggregate": {
      "post": {
        "operationId": "ResultsController_aggregate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AggregateResultDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "results"
        ]
      }
    },
    "/api/v1/results/{userId}/export": {
      "get": {
        "operationId": "ResultsController_export",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cycleId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "format",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "pdf",
                "excel"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "results"
        ]
      }
    },
    "/api/v1/results/{userId}": {
      "get": {
        "operationId": "ResultsController_getDetail",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cycleId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "results"
        ]
      }
    },
...
  },
  "info": {
    "title": "에너지엑스 인사평가 API",
    "description": "모듈러 모놀리식 백엔드 — 응답 봉투 {data}/{data,meta}/{error}",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "OrgSnapshotDto": {
        "type": "object",
        "properties": {
          "group": {
            "type": "string",
            "nullable": true
          },
          "division": {
            "type": "string",
            "nullable": true
          },
          "team": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "group",
          "division",
          "team"
        ]
      },
      "CompareRuleSummaryDto": {
        "type": "object",
        "properties": {
          "competencyIncluded": {
            "type": "boolean",
            "description": "역량 점수가 등급/연봉 산정에 반영됐는지(false=참고용)."
          },
          "gradeScaleLabel": {
            "type": "string",
            "description": "등급 스케일 짧은 라벨(\"S96·A91·B85·C80·D<80\")."
          },
          "source": {
            "type": "string",
            "description": "데이터 출처 — import(과거 엑셀) / aggregate(현행 산정)."
          }
        },
        "required": [
          "competencyIncluded",
          "gradeScaleLabel",
          "source"
        ]
      },
      "CompareTimelineEntryDto": {
        "type": "object",
        "properties": {
          "cycleId": {
            "type": "string"
          },
          "cycleName": {
            "type": "string",
            "nullable": true
          },
          "year": {
            "type": "number",
            "nullable": true
          },
          "finalGrade": {
            "type": "string",
            "nullable": true,
            "enum": [
              "S",
              "A",
              "B",
              "C",
              "D"
            ]
          },
          "finalScore": {
            "type": "number",
            "nullable": true
          },
          "percentile": {
            "type": "number",
            "nullable": true
          },
          "perf": {
            "type": "number",
            "nullable": true,
            "description": "실적 원형."
          },
          "comp": {
            "type": "number",
            "nullable": true,
            "description": "역량 원형(참고용, null 가능)."
          },
          "org": {
            "$ref": "#/components/schemas/OrgSnapshotDto"
          },
          "ruleSummary": {
            "$ref": "#/components/schemas/CompareRuleSummaryDto"
          }
        },
        "required": [
          "cycleId",
          "cycleName",
          "year",
          "finalGrade",
          "finalScore",
          "percentile",
          "perf",
          "comp",
          "org",
          "ruleSummary"
        ]
      },
      "CompareResultDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "userName": {
            "type": "string"
          },
          "employmentStatus": {
            "type": "string",
            "description": "재직 상태(active·resigned 등)."
          },
          "legalEntity": {
            "type": "string",
            "description": "법인(energyx·mirae_plan 등)."
          },
          "timeline": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompareTimelineEntryDto"
            }
          }
        },
        "required": [
          "userId",
          "userName",
          "employmentStatus",
          "legalEntity",
          "timeline"
        ]
      },
      "GradeCountMapDto": {
        "type": "object",
        "properties": {
          "S": {
            "type": "number"
          },
          "A": {
            "type": "number"
          },
          "B": {
            "type": "number"
          },
          "C": {
            "type": "number"
          },
          "D": {
            "type": "number"
          }
        },
        "required": [
          "S",
          "A",
          "B",
          "C",
          "D"
        ]
      },
      "DistributionBucketDto": {
        "type": "object",
        "properties": {
          "deptName": {
            "type": "string",
            "description": "스냅샷/현재 조직명."
          },
          "total": {
            "type": "number"
          },
          "counts": {
            "$ref": "#/components/schemas/GradeCountMapDto"
          },
          "ratios": {
            "description": "등급 비율(%, 소수1).",
            "allOf": [
              {
                "$ref": "#/components/schemas/GradeCountMapDto"
              }
            ]
          }
        },
        "required": [
          "deptName",
          "total",
          "counts",
          "ratios"
        ]
      },
      "DistributionOverallDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "counts": {
            "$ref": "#/components/schemas/GradeCountMapDto"
          },
          "ratios": {
            "$ref": "#/components/schemas/GradeCountMapDto"
          }
        },
        "required": [
          "total",
          "counts",
          "ratios"
        ]
      },
      "DistributionCycleDto": {
        "type": "object",
        "properties": {
          "cycleId": {
            "type": "string"
          },
          "cycleName": {
            "type": "string",
            "nullable": true
          },
          "year": {
            "type": "number",
            "nullable": true
          },
          "buckets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DistributionBucketDto"
            }
          },
          "overall": {
            "$ref": "#/components/schemas/DistributionOverallDto"
          }
        },
        "required": [
          "cycleId",
          "cycleName",
          "year",
          "buckets",
          "overall"
        ]
      },
      "DistributionResultDto": {
        "type": "object",
        "properties": {
          "scope": {
            "type": "string",
            "description": "조직 단위(group·division·team).",
            "enum": [
              "group",
              "division",
              "team"
            ]
          },
          "cycles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DistributionCycleDto"
            }
          }
        },
        "required": [
          "scope",
          "cycles"
        ]
      },
      "SummaryStageDto": {
        "type": "object",
        "properties": {
          "perf": {
            "type": "number",
            "nullable": true
          },
          "comp": {
            "type": "number",
            "nullable": true
          }
        },
        "required": [
          "perf",
          "comp"
        ]
      },
      "SummaryRowDto": {
        "type": "object",
        "properties": {
          "no": {
            "type": "number"
          },
          "userId": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "group": {
            "type": "string",
            "nullable": true
          },
          "division": {
            "type": "string",
            "nullable": true
          },
          "team": {
            "type": "string",
            "nullable": true
          },
          "position": {
            "type": "string",
            "nullable": true,
            "description": "직책 코드(PositionDef.code)."
          },
          "role": {
            "type": "string",
            "nullable": true,
            "description": "역할(hr_admin·division_head·team_lead·employee)."
          },
          "stage1": {
            "$ref": "#/components/schemas/SummaryStageDto"
          },
          "stage2": {
            "$ref": "#/components/schemas/SummaryStageDto"
          },
          "stageFinal": {
            "$ref": "#/components/schemas/SummaryStageDto"
          },
          "sum": {
            "$ref": "#/components/schemas/SummaryStageDto"
          },
          "finalScore": {
            "type": "number",
            "nullable": true
          },
          "finalGrade": {
            "type": "string",
            "nullable": true,
            "description": "최종 등급(S~D)."
          },
          "source": {
            "type": "string",
            "description": "데이터 출처 — import(과거 엑셀) / live(현행 산정).",
            "enum": [
              "import",
              "live"
            ]
          }
        },
        "required": [
          "no",
          "userId",
          "name",
          "group",
          "division",
          "team",
          "position",
          "role",
          "stage1",
          "stage2",
          "stageFinal",
          "sum",
          "finalScore",
          "finalGrade",
          "source"
        ]
      },
      "AggregateResultDto": {
        "type": "object",
        "properties": {
          "cycleId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          }
        },
        "required": [
          "cycleId",
          "userId"
        ]
      },
...
    }
  }
}