# packages/contracts/src/generated/model/distributionCycleDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * 에너지엑스 인사평가 API
 * 모듈러 모놀리식 백엔드 — 응답 봉투 {data}/{data,meta}/{error}
 * OpenAPI spec version: 1.0.0
 */
import type { DistributionBucketDto } from './distributionBucketDto';
import type { DistributionOverallDto } from './distributionOverallDto';

export interface DistributionCycleDto {
  cycleId: string;
  /** @nullable */
  cycleName: string | null;
  /** @nullable */
  year: number | null;
  buckets: DistributionBucketDto[];
  overall: DistributionOverallDto;
}